Rename platform -> standard_platform
